+++ lib/ajax_pagination/controller_additions.rb
@@ -2,33 +2,34 @@ module AjaxPagination
   # This module is automatically added to all controllers
   module ControllerAdditions
     module ClassMethods
-      # Adds default render behaviour for requests with a pagination parameter matching a certain name. By default, this name is the empty string. However, it can be changed.
+      # Adds default render behaviour for AJAX requests for a section with matching a certain name. By default, this name is the empty string. However, it can be changed.
       # Options:
+      # [:+section_id+]
+      #   The AJAX section name which should be matched to invoke AJAX Pagination response. Defaults to "global".
+      #
       # [:+pagination+]
-      #   The pagination name which should be matched to invoke AJAX Pagination response. Defaults to the empty string "". This is unlike other
-      #   methods which default to "page". However, this is a better default in this case because this affects more than one controller/action.
+      #   Deprecated. Alias for section_id.
       #
       # [:+render+]
-      #   Overrides default render behaviour for AJAX Pagination, which is to render the partial with name matching the pagination option,
+      #   Overrides default render behaviour for AJAX Pagination, which is to render the partial with name matching the section_id option,
       #   or if it does not exist, renders the default template
       #
       def ajax_respond(options = {});
         # instead of defining default render normally, we save an unbound reference to original function in case it was already defined, since we want to retain the original behaviour, and add to it (if the method is redefined after, this new behaviour is lost, but at least we don't remove others' behaviour - note that this also allows multiple invocations of this with different parameters)
         default_render = self.instance_method(:default_render) # get a reference to original method
-        pagination = options[:pagination] || ""
+        section_id = options[:section_id] || options[:pagination] || "global"
         view = options[:render] || nil
         define_method(:default_render) do |*args|
-          paramspagination = request.GET[:pagination] || params[:pagination]
-          if paramspagination && paramspagination == pagination && request.format == "html" # override if calling AJAX Pagination
+          if ajax_section && ajax_section == section_id && request.format == "html" # override if calling AJAX Pagination
             unless view
-              if lookup_context.find_all("#{params[:controller]}/_#{paramspagination}").any?
-                view = { :partial => paramspagination } # render partial, layout is off
+              if lookup_context.find_all("#{params[:controller]}/_#{ajax_section}").any?
+                view = { :partial => ajax_section } # render partial, layout is off
               else
                 view = { :layout => false } # render default view, but turn off layout
               end
             end
             respond_to do |format|
-              ajax_respond format, :pagination => pagination, :render => view
+              ajax_respond format, :section_id => section_id, :render => view
             end
           else # otherwise do what would have been done
             default_render.bind(self).call(*args) # call original method of the same name
@@ -38,7 +39,23 @@ module AjaxPagination
     end
     def self.included(base)
       base.extend ClassMethods
+
+      if AjaxPagination.redirect_after_filter == true
+        base.after_filter :ajax_pagination_redirect
+      end
+
+      base.before_filter do
+        @_ajax_section = request.GET[:pagination] || params[:pagination]
+        params.delete(:pagination) if request.get?
+      end
     end
+
+    # Returns the name of the ajax section to be responded to, if present. Otherwise this is not an AJAX Pagination request
+    # (ie. this is otherwise a normal full page request). When no specific ajax section is requested, returns nil.
+    def ajax_section
+      @_ajax_section
+    end
+
     # Registers an ajax response in html format when params [:pagination] matches options [:pagination] ( = "page" by default).
     # AJAX Pagination uses this response to render only the content which has changed. When this format is triggered,
     # a partial is passed back, and sent to AJAX Pagination as a function argument in javascript.
@@ -56,9 +73,12 @@ module AjaxPagination
     #   end
     #
     # Options:
+    # [:+section_id+]
+    #   Changes the AJAX section name triggering this response. Triggered when ajax_section == options [:section_id].
+    #   Defaults to "global"
+    #
     # [:+pagination+]
-    #   Changes the pagination name triggering this response. Triggered when params [:pagination] == options [:pagination].
-    #   Defaults to "page"
+    #   Deprecated. Alias for section_id
     #
     # [:+render+]
     #   Changes the default template/partial that is rendered by this response. The value can be any object,
@@ -75,12 +95,11 @@ module AjaxPagination
     #     end
     #
     def ajax_respond(format,options = {})
-      paramspagination = request.GET[:pagination] || params[:pagination]
-      if paramspagination == (options[:pagination] || 'page').to_s
+      if ajax_section == (options[:section_id] || options[:pagination] || 'global').to_s
         if options[:render]
           view = options[:render] # render non partial
-        elsif lookup_context.find_all(params[:controller] + "/_" + paramspagination).any?
-          view = {:partial => paramspagination} # render partial of the same name as pagination
+        elsif lookup_context.find_all(params[:controller] + "/_" + ajax_section).any?
+          view = {:partial => ajax_section} # render partial of the same name as pagination
         else # render usual view
           view = {}
         end
@@ -94,8 +113,8 @@ module AjaxPagination
 
     # Tests whether an AJAX Pagination partial might be displayed in the view. If the response is not directly controlled by
     # AJAX Pagination, it will return true, because the partial might be displayed. The response is handled by AJAX Pagination
-    # if the format is html, and the params [:pagination] parameter is set. If it is set, then it will return whether
-    # params [:pagination] == pagination (the name of the pagination set, which defaults to page).
+    # if the format is html, and the ajax_section parameter is set. If it is set, then it will return whether
+    # ajax_section == section_id (the name of the section, which defaults to page).
     #
     # This method is a convenience function so that the controller does not need to perform heavy computation which might only
     # be required if a certain pagination partial is displayed.
@@ -126,9 +145,8 @@ module AjaxPagination
     #   end
     #
     # The heavy computation will only be performed on posts which will be displayed when AJAX Pagination only wants a partial.
-    def ajax_section_displayed?(pagination = :page)
-      paramspagination = request.GET[:pagination] || params[:pagination]
-      (!request.format.html?) || (paramspagination.nil?) || (paramspagination == pagination.to_s)
+    def ajax_section_displayed?(section_id = :global)
+      (ajax_section.nil?) || (ajax_section == section_id.to_s)
     end
 
     # This after_filter method is automatically included by AJAX Paginate, and does not need to be included manually. However,
@@ -144,8 +162,7 @@ module AjaxPagination
     # This filter should not affect other uses, because only AJAX calls trigger this. In addition, a ?pagination= parameter is required.
     # Therefore other AJAX libraries or usage otherwise should not be affected.
     def ajax_pagination_redirect
-      paramspagination = request.GET[:pagination] || params[:pagination]
-      if request.xhr? && paramspagination && response.status==302 # alter redirect response so that it can be detected by the client javascript
+      if request.xhr? && ajax_section && response.status==302 # alter redirect response so that it can be detected by the client javascript
         response.status = 200 # change response to OK, location header is preserved, so AJAX can get the new page manually
       end
     end