+++ spec/ajax_pagination/integration/nojavascript_spec.rb
@@ -15,15 +15,15 @@ describe 'paginating without javascript', :type => :request, :driver => :rack_te
   it 'still paginates nested and multiple paginated sections' do
     visit(changelog_url)
     page.should have_selector('.previous_page.disabled')
-    find('#_paginated_section').find('.next_page').click
+    find('#page').find('.next_page').click
     page.should have_no_selector('.previous_page.disabled')
     find('#signin').click
     visit(posts_url)
-    page.should have_selector('#upcomingpage_paginated_section .previous_page.disabled')
-    find('#upcomingpage_paginated_section').find('.next_page').click
-    page.should have_no_selector('#upcomingpage_paginated_section .previous_page.disabled')
-    page.should have_selector('#page_paginated_section .previous_page.disabled')
-    find('#page_paginated_section').find('.next_page').click
-    page.should have_no_selector('#page_paginated_section .previous_page.disabled')
+    page.should have_selector('#upcomingpage .previous_page.disabled')
+    find('#upcomingpage').find('.next_page').click
+    page.should have_no_selector('#upcomingpage .previous_page.disabled')
+    page.should have_selector('#page .previous_page.disabled')
+    find('#page').find('.next_page').click
+    page.should have_no_selector('#page .previous_page.disabled')
   end
 end